+++ src/storage/mvcc/reader/reader.rs
@@ -588,11 +588,13 @@ mod tests {
                         let handle = rocksdb_util::get_cf_handle(db, cf).unwrap();
                         wb.delete_cf(handle, &k).unwrap();
                     }
-                    Modify::DeleteRange(cf, k1, k2) => {
-                        let k1 = keys::data_key(k1.as_encoded());
-                        let k2 = keys::data_key(k2.as_encoded());
-                        let handle = rocksdb_util::get_cf_handle(db, cf).unwrap();
-                        wb.delete_range_cf(handle, &k1, &k2).unwrap();
+                    Modify::DeleteRange(cf, k1, k2, is_unsafe) => {
+                        if !is_unsafe {
+                            let k1 = keys::data_key(k1.as_encoded());
+                            let k2 = keys::data_key(k2.as_encoded());
+                            let handle = rocksdb_util::get_cf_handle(db, cf).unwrap();
+                            wb.delete_range_cf(handle, &k1, &k2).unwrap();
+                        }
                     }
                 }
             }
